// assets/app/Auth/types/EndSessionsRequest.ts
// Code generated by tsgen. DO NOT EDIT.

import { isApiFieldErrors } from '@/shared/Fetch/Envelope/ApiFieldErrors';
import type { ApiMessage } from '@/shared/Fetch/Envelope/ApiMessage';

export type EndSessionsRequest = {
    ids: string[];
    all: boolean;
    ip: string;
};

export type EndSessionsRequestErrors = {
    general?: ApiMessage;
    ids?: ApiMessage;
    [key: `ids[${number}]`]: ApiMessage;
    all?: ApiMessage;
    ip?: ApiMessage;
};

export const isEndSessionsRequestErrors = (v: unknown): v is EndSessionsRequestErrors =>
    isApiFieldErrors(/^(?:general|ids|ids\[\d+\]|all|ip)$/, v);